feat: use external component for footer
